app/requirements.txt
anyio==3.7.1
appdirs==1.4.4
asyncer==0.0.2
idna==3.4
lxml==4.9.3
nats-py==2.6.0
packaging==23.2
py4j==0.10.9.7
pyshark==0.6
pyspark==3.5.0
sniffio==1.3.0
termcolor==2.3.0
